requirements.dev.txt
Django==2.0.7
django-environ==0.4.5
django-filter==1.1.0
django-extensions==2.1.0
pytz==2018.5
requests==2.19.1
djangorestframework==3.8.2
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
psycopg2==2.7
django-cors-headers==2.3.0
dry-rest-permissions==0.1.10
drf-oidc-auth==0.9
django-queryinspect==1.0.0
django-cors-headers==2.3.0
gitdb==0.6.4
nose==1.3.7
mock==2.0.0
gitdb2==2.0.5
PyYAML==5.1
GitPython==2.1.11
json2html==1.2.1
celery==4.2.1
redis==3.2.0
rethinkdb==2.3.0.post6
docker==3.1.3
docker-pycreds==0.3.0
boto3